requirements.txt
# Usage: pip install -r requirements.txt
tensorflow
termcolor
pybase64
pillow
streamlit
colorama
